feat: functions has been improved
feat: added sitemap.xml generator
fix: lint issues
