+++ functions/src/redirection-record-model.ts
@@ -1,0 +1,7 @@
+/** Redirection Record Model */
+export class RedirectionRecordModel {
+    /** http code */
+    code?: number;
+    /** url to redirect */
+    url?: string;
+}
