+++ functions/src/site-map-url-model.ts
@@ -1,0 +1,11 @@
+/** Site Map Url Model */
+export class SiteMapUrlModel {
+    /** url of page */
+    url: string;
+    /** changefreq of page */
+    changefreq?: string;
+    /** priority of page */
+    priority?: number;
+    /** image url of page */
+    img?: string;
+}
