+++ migrate-to-firebase/src/download-data-from-firestore.ts
@@ -7,7 +7,7 @@ const serviceAccount = require('../supermurat-com-service-key.json');
 
 admin.initializeApp({
     credential: admin.credential.cert(serviceAccount),
-    databaseURL: `https://${serviceAccount.project_id}.firebaseio.com`,
+    databaseURL: `https://${serviceAccount.project_id}.firebaseio.com`
 });
 admin.app().firestore().settings({timestampsInSnapshots: true});
 const db = admin.firestore();
